LC-maximize-happiness-of-selected-children [Runtime: 155 ms; Memory: 107.9 MB]

+++ greedy/maximize-happiness-of-selected-children.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long maximumHappinessSum(vector<int>& happiness, int k) {
-        int temp=0,n=happiness.size(),ans=0;
+        long long temp=0,n=happiness.size(),ans=0;
         sort(happiness.begin(),happiness.end(),greater<>());
         for(int i=0; i<k && temp<k; i++){
             if(happiness[i]-temp<=0) break;
